:truck: api: move cancel certif controller to src
